Crate arle

ACLE in Rust

Developer notes

Below is a list of built-in targets that are representative of the different ARM architectures; the list includes the target_features they possess.

armv4t-unknown-linux-gnueabi - ARMv4 - +v4t

armv5te-unknown-linux-gnueabi - ARMv5TE - +v4t +v5te

arm-unknown-linux-gnueabi - ARMv6 - +v4t +v5te +v6

thumbv6m-none-eabi - ARMv6-M - +v4t +v5te +v6 +thumb-mode +mclass

armv7-unknown-linux-gnueabihf - ARMv7-A - +v4t +v5te +v6 +v6k +v6t2 +v7 +dsp +thumb2 +aclass

armv7r-none-eabi - ARMv7-R - +v4t +v5te +v6 +v6k +v6t2 +v7 +dsp +thumb2 +rclass

thumbv7m-none-eabi - ARMv7-M - +v4t +v5te +v6 +v6k +v6t2 +v7 +thumb2 +thumb-mode +mclass

thumbv7em-none-eabi - ARMv7E-M - +v4t +v5te +v6 +v6k +v6t2 +v7 +dsp +thumb2 +thumb-mode +mclass

aarch64-unknown-linux-gnu - ARMv8-A (AArch64) - +fp +neon

From section 10.1 of ACLE:

"In the sequence of Arm architectures { v5, v5TE, v6, v6T2, v7 } each architecture includes its predecessor instruction set."

"In the sequence of Thumb-only architectures { v6-M, v7-M, v7E-M } each architecture includes its predecessor instruction set."

From that info and from looking at how LLVM features work (using custom targets) we can identify features that are subsets of others:

Legend: a < b reads as "a is a subset of b"; this means that if b is enabled then a is enabled as well.

v4t < v5te < v6 < v6k < v6t2 < v7 < v8

v6 < v8m < v6t2

v7 < v8m.main

References

ACLE Q2 2018

Structs

APSR
Application Program Status Register

ISH
Inner Shareable is the required shareability domain, reads and writes are the required access types

ISHST
Inner Shareable is the required shareability domain, writes are the required access type

NSH
Non-shareable is the required shareability domain, reads and writes are the required access types

NSHST
Non-shareable is the required shareability domain, writes are the required access type

OSH
Outer Shareable is the required shareability domain, reads and writes are the required access types

OSHST
Outer Shareable is the required shareability domain, writes are the required access type

ST
Full system is the required shareability domain, writes are the required access type

SY
Full system is the required shareability domain, reads and writes are the required access types

Functions

__dbg⚠
Generates a DBG instruction.

__dmb⚠
Generates a DMB (data memory barrier) instruction or equivalent CP15 instruction.

__dsb⚠
Generates a DSB (data synchronization barrier) instruction or equivalent CP15 instruction.

__isb⚠
Generates an ISB (instruction synchronization barrier) instruction or equivalent CP15 instruction.

__nop⚠
Generates an unspecified no-op instruction.

__rsr⚠
Reads a 32-bit system register

__rsrp⚠
Reads a system register containing an address

__sev⚠
Generates a SEV (send a global event) hint instruction.

__wfe⚠
Generates a WFE (wait for event) hint instruction, or nothing.

__wfi⚠
Generates a WFI (wait for interrupt) hint instruction, or nothing.

__wsr⚠
Writes a 32-bit system register

__wsrp⚠
Writes a system register containing an address

__yield⚠
Generates a YIELD hint instruction.